Add mockup directory.
The mockup is designed for putting sparkles demo to snoek.
For example, the mockup/interview-timeslot.ots is the
libreoffice template we are using to track employee
interview schedule and status. It is good if we implement
such kind of tool in snoek so that people is able to track
and share these stuff online.